Get Appointment Slots from MCM
Although this is not a complete implementation, we have added enough
scaffolding to the code to be able to request _some_ available
appointments from the MCM Scheduling system.

We have added Flurl as a dependency in order to have a fluent HttpClient
with which we can also write some useful tests.

A lot of the values are currently hardcoded, this is to get things
moving, any shortcuts have been marked with "TODO" and will be addressed
soon.