Add job codes mapper
In order to produce the trade codes needed by the MCM system we have
added a `JobCodesMapper` which, given a SoR code, will interrogate a
provided json file to find the associated trade codes. There appears to
be a rule for mapping from SoR codes to trade codes, but instead of
making a mistake/having to handle edge cases, it is easier to simply
include a json file with the mappings.