Derive _some_ slot descriptions for MCM
MCM requires a slot description such as "10:00-12:00" to identify an
appointment. These are nice and predictable and can therefore be derived
from the slot start time and end time. There are other slot descriptions
such as "Sat AM" that this commit doesn't cover, and will cause errors,
however the assumption is that these appointments _shouldn't_ be
returned by the MCM system in production so we will need to keep an eye
on this.