Remove "named" slots
We are removing slots such as "Sat AM" and "ALL" from the list of
available slots shown to the user. This may not be necessary in
production as these slots _should_ not be returned from MCM, but we
don't necessarily want to rely on this behaviour as it is not something
that we control.